Fix SPI bugs #9 (NULL platform_ops) and #10 (missing CS toggle), widen chip_select to uint16_t
Bug #9: Both TX and RX SPI init params had platform_ops = NULL, causing
adf4382_init() -> no_os_spi_init() to fail with -EINVAL. Fixed by setting
platform_ops = &stm32_spi_ops and passing stm32_spi_extra with correct CS
port/pin for each device.

Bug #10: stm32_spi_write_and_read() never toggled chip select. Since TX
and RX ADF4382A share SPI4, every register write hit both PLLs. Rewrote
stm32_spi.c to assert CS LOW before transfer and deassert HIGH after,
using stm32_spi_extra metadata. Backward-compatible: legacy callers
(e.g., AD9523) with cs_port=NULL skip CS management.

Also widened chip_select from uint8_t to uint16_t in no_os_spi.h since
STM32 GPIO_PIN_xx values (e.g., GPIO_PIN_14=0x4000) overflow uint8_t.

10/10 tests pass (8 original + 2 new regression tests).

+++ 9_Firmware/9_1_Microcontroller/tests/ad_driver_mock.h
@@ -47,7 +47,7 @@ struct no_os_spi_platform_ops {
 struct no_os_spi_init_param {
     uint32_t                            device_id;
     uint32_t                            max_speed_hz;
-    uint8_t                             chip_select;
+    uint16_t                            chip_select;
     enum no_os_spi_mode                 mode;
     enum no_os_spi_bit_order            bit_order;
     enum no_os_spi_lanes                lanes;
